Rename module to match hermes.error
