+++ src/hermes/model/error.py
